updating hmac verification function
Shopify recently added session to their request parameters when logging in from the the app page of the shopify admin. So with the old code anything not hardcoded in the old if statement would be getting stripped from the hmac verification and according to we should only be removing the hmac from the request and nothing else included docs for verification. 

https://help.shopify.com/en/api/getting-started/authentication/oauth#verificatio