Throw an exception when cart clear request fails
When a Shopify shop has password enabled Shopify replies with status
code 200 and an html page requesting the password. When this happens
we get a null value when json_decode fails. This is an invalid cart
so we throw a pretty generic ShopifyException so the caller can handle
the error.